moved class files and modules to dir auto-reporting/
