+++ auto-reporting/credentials.rb
