+++ auto-reporting/csv_writer.rb
