+++ auto-reporting/quickstart.rb
